feat: add search route and controller wip:add Broadcast events and make them work

+++ database/migrations/2022_06_28_194225_create_movies_table.php
@@ -21,7 +21,7 @@ return new class extends Migration {
             $table->integer('release_date');
             $table->json('description');
             $table->json('director');
-            $table->float('income');
+            $table->bigInteger('income');
             $table->timestamps();
         });
     }
